requirements.txt
flask
google-genai
selenium
webdriver-manager
pytest
